Use resourceful route for user status
