+++ config/routes.rb
@@ -269,11 +269,11 @@ OpenStreetMap::Application.routes.draw do
   post "/diary_comments/:comment/unhide" => "diary_comments#unhide", :comment => /\d+/, :as => :unhide_diary_comment
 
   # user pages
-  resources :users, :path => "user", :param => :display_name, :only => [:show, :destroy] do
+  resources :users, :path => "user", :param => :display_name, :only => :show do
     resource :role, :controller => "user_roles", :path => "roles/:role", :only => [:create, :destroy]
+    resource :status, :module => :users, :only => [:update, :destroy]
   end
   get "/user/:display_name/account", :to => redirect(:path => "/account/edit")
-  post "/user/:display_name/set_status" => "users#set_status", :as => :set_status_user
 
   resource :account, :only => [:edit, :update, :destroy]
